Refactor of CRC cleaned up code and removed uneeded imports, added override when needed.

+++ edu.harvard.i2b2.crc/src/server/edu/harvard/i2b2/crc/dao/setfinder/GaussianBoxMuller.java
@@ -13,7 +13,7 @@ public class GaussianBoxMuller {
 			obNum = Math.round(boxMuller(true, rangeValue, 0));
 		}
 		// if less than three, then return 0
-		return (double) originalValue + obNum;
+		return originalValue + obNum;
 
 	}
